Created a new class, StringUtils. Motivation: To be less space-big as Apache-Commons StringUtils, while still having key methods.
