+++ app/src/main/java/com/niupiao/niupiao/fragments/my_tickets/TicketFragment.java
@@ -23,11 +23,10 @@ import com.niupiao.niupiao.activities.TicketActivity;
 import com.niupiao.niupiao.models.Event;
 import com.niupiao.niupiao.models.Ticket;
 import com.niupiao.niupiao.models.User;
+import com.niupiao.niupiao.utils.StringUtils;
 
 import org.joda.time.LocalDate;
 import org.joda.time.LocalTime;
-import org.joda.time.format.DateTimeFormat;
-import org.joda.time.format.DateTimeFormatter;
 
 /**
  * Created by Inanity on 2/27/15.
@@ -71,7 +70,7 @@ public class TicketFragment extends Fragment {
         //Set up format for embeddedInfo
         String format = "%s ";
         int numArgs = 17;
-        format = TicketFragment.repeatHelper(format, numArgs);
+        format = StringUtils.repeatHelper(format, numArgs);
         format = format.substring(0, format.length()-1);
 
         String embeddedInfo = String.format(format,
@@ -82,7 +81,6 @@ public class TicketFragment extends Fragment {
                 "Current Local Time:",
                 localTime.getHourOfDay() + ":" + localTime.getMinuteOfHour(),
                 "\n",
-
 
                 "Event Name:",
                 event.getName(),
@@ -135,21 +133,5 @@ public class TicketFragment extends Fragment {
             Toast.makeText(getActivity().getApplicationContext(), "Error with QR Code Generation",
                     Toast.LENGTH_SHORT).show();
         }
-    }
-
-    /*
-     * A very small utility method. Based in part on producing the same output as
-     * a similar method as org.apache.commons.lang3.StringUtils
-     *
-     * @param String str is the string to be repeated.
-     * @param int num is the number of times to repeat the string. If 0, an empty string is returned.
-     *        Values lss than 0 produce output identical to 0.
-     */
-    public static String repeatHelper(String str, int num){
-        String output = "";
-        for(int i = 0; i < num; i ++){
-            output = output + str;
-        }
-        return output;
     }
 }
